+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 2015/04/19: [#28](https://github.com/andrewvy/slack-pongbot/pull/28) - Accept both challenge `single` and `singles`, `double` and `doubles` - [@dblock](https://github.com/dblock).
 * 2015/04/17: [#19](https://github.com/andrewvy/slack-pongbot/pull/19) - Use jshint for code linting - [@dblock](https://github.com/dblock).
 * 2015/04/14: [#13](https://github.com/andrewvy/slack-pongbot/pull/13) - Fix: `undefined variable err` in `getEveryone` - [@dblock](https://github.com/dblock).
 * 2015/04/14: [#12](https://github.com/andrewvy/slack-pongbot/pull/12) - Added tests - [@dblock](https://github.com/dblock).